NEW: added example configuration file to deal with static windows on a big screen
the applications can be set by the standard arguments for the WindowManager object plus a new argument "init_exe_sequence". this argument will be evaluated to determine the user's wishes concerning the applications to be filled into the window spots and a possible further string value which might be tested for containment within the specific application title. this makes sense if there are more than one application with the same execution path but different titles.